vid_social_api/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django-cors-headers==4.3.1
PyYAML==6.0.1
requests
arrow
moviepy
channels==4.0.0
celery==5.3.6
agora-token-builder==1.0.0
openpyxl==3.1.2
django-extensions==3.2.3
# sudo apt-get install python3-dev default-libmysqlclient-dev build-essential
# mysqlclient==2.2.0